- ACPI: Don't redefine standard DDK types - TCPIP/IP/TDI/Uniata: Don't use user mode types - USBDRIVER: Remove a peculiar redefinition of ULONG_PTR - NDIS: Remove duplicate (and undocumented) types from ndis.h, defined in an internal NDIS header long ago. Also fix some warnings, define UINT as appropriate, and add a clever (not really) workaround for a gcc/msvc incompatibility. -- For some obnoxious reason msvc is fine with redeclaring types compatibly, while gcc treats it as an error. PSDK and DDK require this, so we need...extended header checks.
svn path=/trunk/; revision=37688

+++ reactos/drivers/network/tcpip/tcpip/iinfo.c
@@ -83,8 +83,8 @@ TDI_STATUS InfoTdiQueryGetArptableMIB(TDIEntityID *ID,
 				      PNDIS_BUFFER Buffer,
 				      PUINT BufferSize) {
     NTSTATUS Status;
-    DWORD NumNeighbors = NBCopyNeighbors( Interface, NULL );
-    DWORD MemSize = NumNeighbors * sizeof(IPARP_ENTRY);
+    ULONG NumNeighbors = NBCopyNeighbors( Interface, NULL );
+    ULONG MemSize = NumNeighbors * sizeof(IPARP_ENTRY);
     PIPARP_ENTRY ArpEntries =
 	ExAllocatePoolWithTag
 	( NonPagedPool, MemSize, FOURCC('A','R','P','t') );
@@ -104,7 +104,7 @@ TDI_STATUS InfoTdiQueryGetArpCapability(TDIEntityID *ID,
 					PNDIS_BUFFER Buffer,
 					PUINT BufferSize) {
     NTSTATUS Status;
-    DWORD Capability = 0x280;
+    ULONG Capability = 0x280;
 
     TI_DbgPrint(MID_TRACE,("Copying out %d bytes (AT_ENTITY capability)\n",
 			   sizeof(Capability)));
